Remove AST mutation from comprehension macros.
We still had a few shady argument-modifying macros sticking around
with icollect, collect, and accumulate.

Also as a bonus port all the loop tests over to using an == macro
instead of string eval.

Copying for macro args doesn't need to be deep copying; single-level
is fine.

Some of the copying in the arrow macros weren't necessary; remove them.