Observe DESTDIR environment variable in Makefile
This commit modifies the `install` target of the Makefile so that it
installs Fennel into the directory specified by the DESTDIR environment
variable, if that is set. This allows Fennel to be installed into an
local directory tree mirroring a root filesystem, which is needed for
building a binary Debian package.