0.4.x: Remove compiled Lua artefacts from source tree
Since the files fennelfriend.lua and fennelview.lua are regenerated
automatically from their corresponding Fennel sources by the Makefile if
not present, this commit removes them from the source tree. For the
Debian packaging of Fennel, this alleviates the need for a downstream
patch to remove them at build time; the Debian policy does not allow
pre-compiled artefacts in source packages.