Fix a bug emitting table literals modified by macros.
What happened is that the parser puts a `keys` field in a metatable on
every key/value table literal it reads. This allows the table to be
reconstructed exactly as it was originally in the input; even though
key/value tables are un-ordered in their representation, we can emit
them with their keys in the original order.

The problem is that a macro can modify a table in a way that causes it
to get out of sync with `keys` table. (Basically any normal table
modification in a macro that happens to a literal table does this.)
When the compiler emitted a Lua table, it only paid attention to `keys`
and if it disagreed with the actual value of the table itself, the
`keys` table won.

The correct behavior is to use `keys` only for ordering: we must
ignore entries which don't exist in the table, and we must ensure that
legitimate keys which aren't found in `keys` are still emitted in the Lua.

The old code also appeared to check for duplicates in the `keys` table
which as far as I can tell never actually happens.

Bonus: fix up the macro tests to use the == macro more consistently.