Go back to the default package.path to find luaunit.
Previously we adjusted this so it loaded luaunit as just "luaunit" but
that meant that you couldn't just run `lua test/init.lua` since you
had to set the package path. This was done in order to make debian
packaging easier, but they can work around the problem in the
packaging scripts; it's not worth the headache in day to day dev.