Set filename on import-macros internal expression for relative requires
The module names passed to import-macros may be a string or an
expression. Expressions are evaluated and the module name and filename
of the module that called import-macros are passed in as arguments.

Internally import-macros calls require-macros with a generated
expression, we have to copy the correct filename over to our internal
expression before the module names are resolved.