fix pretty printing sequences containing comments
Previously:
```
(print [:really-long-string-to-put-this-on-multiple-lines-asdfasdfasdf
        ;; comment
])

(print [:foo ;; comment
        :bar])
```

would format to:
```
(print [:really-long-string-to-put-this-on-multiple-lines-asdfasdfasdf
        ;; comment])

(print [:foo ;; comment :bar])
```

but now it formats to:
```
(print [:really-long-string-to-put-this-on-multiple-lines-asdfasdfasdf
        ;; comment
        ])

(print [:foo
        ;; comment
        :bar])
```

The approach taken computes whether an ast node is a comment within fennelview,
as opposed to doing this in the parser.

Handling a comment that is *not* the last item in a sequence is done by setting
`multiline?` to true if any item is a comment.

Handling a comment that is the last item in a sequence is a special case since
we need to put a newline (and indent) after it.

AFAICT, there are two possible places to adapt fennelview to handle this:
  1) when printing the sequence closing delimiter (i.e. adding a new line before
     the closing delimiter in `pp-sequence`)
  2) when printing the last comment (i.e. appending a `""` to `c` in
     `comment-view`)

The challenge with (2) is that we need some context in `comment-view` to know if
it's the last item in a sequence, which requires drilling a flag through a call
stack that looks like this:
    - pp-sequence
    - pp
    - pp-table
    - pp-metamethod
    - comment-view

Perhaps this could be avoided by mutating `options`, but then we'd need to worry
about cleaning up after the call to `comment-view`.

This patch implements (1), which feels like a cleaner change.