Use column number information from parser in error messages.
This introduces a utils.len function which will give the length of a
string in characters when that is available (lua 5.3+ or older luas
with a backport of the utf8 library, such as love2d) but falls back to
bytes otherwise. This means that the pinpoint ------^ feature in error
messages will make a best-effort to calculate the offsets correctly
but can sometimes be wrong. Even when the utf8 library is available,
it cannot reliably tell you the *width* of a string; for
instance (utf8.len "วัด") gives 3, which is the number of codepoints in
the string, but the string is only 2 characters wide for purposes of
calculating the column.

Currently the attempt to be utf8-aware is only done in the error
pinpointing code. The parser column number information is strictly
byte-based.

We are also emitting column numbers in the textual error messages
emitted whether or not fennel.friend is active:

    Compile error in scratch.fnl:4:13

The presence of column information on the AST nodes allows us to
simplify the reading of source files in fennel.friend fairly
significantly because we no longer need to depend on counting as it
steps thru the file to track byte offsets that way.