Add plugins option to fennel.parser
There is a new function `utils.hook-opts` which takes in an `?options`
table, and try to use the plugins there first. It will fall back to
root.options if the `?options` table isn't present or doesn't have a
list of plugins. `fennel.parser` now calls `utils.hook-opts` instead
of `utils.hook`, and provides the options table. When the user adds
plugins to the options table, `utils.hook-opts` will pick up the
user's plugins.

There is also a test case and a mention of the change inside of
`api.md`.