bugfix: use new utils.get-in to look up macro modules in ,doc
The doc command was only doing a straight key lookup on scope.macros,
which does not account for macro modules imported as a table.

For example, `(import-macros m :path.to.macromodule) ,doc m.foo` would
fail to look up the module correctly from the REPL.

This adds a new util function that can likely replace manual code
elsewhere in the compiler to give us a net reduction in code code length
+ complexity.

Also updated `find-macro` the compiler to use utils.get-in. There are
more places where utils.get-in can likely be subbed in, but since we're
close to a release we can hold off on the additional refactoring until
after 1.2.0 is out.