bugfix: assert macro module is table before setting macro-loaded
The `add-macros` util function, which the `macros` special and
`require-macros` (when invoked directly) will use to apply all macros to
the current scope, contains assertion for macros to be a table. However,
`require-macros` won't invoke this when indirectly invoked by
`imort-macros`.

This resulted in an unfriendly erorr message from `import-macros` when
used on a non-table-returning macro module, but additionally meant that,
even in case of error, the macro module would have already been cached
in fennel.macro-loaded.