Fix filename check for tables and sequences
propagate-source-info replaces bad source information with good source
information in the output of a macro. However, the check to see if the
source information is relevant assumes that node is a list, and
mistakenly misreads the filename for [] sequences and {} tables. This
means that good source information could possibly be replaced with
subpar information.

This commit changes propagate-source-info so it reads the filename
from the src information instead of directly from the node.