Make macros copy lists, but not symbols
In order to keep the AST intact, some macros create modified copies of
their arguments instead of operating on the arguments themselves. This
patch limits those copies to only happen on lists, and not symbols. It
doesn't affect the compiler's output, but makes it friendlier for my
fennel-ls project.

The cases where a symbol gets copied are all performing the same
operation: potentially wrap a symbol in a list if it's not already
wrapped in one, and make sure that the list is "fresh" and not the same
as the input. Previously this seems to be done in a couple ways, but
this patch switches all of them to the same idiom:

```fnl
(if (list? e) (copy e) (list e))
```

If the input is a list, create a copy, otherwise create a fresh list
with the *original* `e`. This way, when `e` is a symbol, it shows up
unchanged in the output of the macro.