Fix parsing of AST nodes which immediately follow a comment.
https://todo.sr.ht/~technomancy/fennel/124

This only manifests as a problem if you parse with comments turned on,
which basically only happens in fnlfmt, but it's actually a bug with
Fennel's own parser. The dispatch function was setting
whitespace-since-dispatch to false even when the thing being
dispatched was a comment (in which case it had already parsed the newline).