Multi-phase macros.
Most of the built-in macros are very simple, and writing them in "raw
fennel" without access to any of the built-in macros isn't so
bad. However, the pattern matching implementation is quite complex!
It's a pain to write it without access to when/doto/etc.

This changes it so that after the core macros are loaded, a second
phase is used to load the pattern matching macros.

The match macros themselves have not been changed to use the macros
they newly have access to, except one place to ensure it does in fact
work. But this will allow the implementation of match to be easier to read.