Use stablepairs in the compiler environment for pairs.
Using regular pairs in macros leads to builds which are
nondeterministic; performing a rebuild of exactly the same input can
result in different compiler output.

We take care to use stablepairs in the compiler wherever we're
emitting order-based things based on pairs, because otherwise we get
different results depending on arbitrary factors. But we still had
some macros which used pairs to emit code.

We could expose stablepairs as stablepairs and leave pairs alone,
updating the built-in macros to use stablepairs. But this would mean
that user macros would have to go out of their way to achieve
reproducible builds, whereas we want to encourage reproducible builds
by default. The stablepairs function is slower, but in the context of
macros this shouldn't make much of a difference.

The reason that pairs was made unstable in regular Lua has to do with
avoiding attacks in servers which may receive hostile input that is
used as table keys, but that rationale does not apply to the Fennel
compiler, so it's appropriate to go out of our way to achieve the old
behavior.