Backport all the new macros from 1.2.0.
We're pcall+requiring fennel view from the location it's going to be
when the bootstrap compiler is used on the later versions; it won't be
present in the current place here but that's fine; it should fail
gracefully until macrodebug is called, which it never will be on this branch.

We have to backport the way that require-macros can return the macro
table because that's now how import-macros works.

Disable some tests we just don't care about any more.