Implement faccumulate in terms of accumulate
We've discussed adding `faccumulate` for a while, more or less as soon
as `fcollect` came to be. Personally, I've found its absence more
jarring now that `fcollect` is part of the language, so I figured I'd go
about adding.

It turns out to have a very small footprint; simply needed to swap
`each` out for `for` in the implementation!