match refactoring, (or) patterns, matchless macro
Improves the `match` macro to add support for new patterns. For a long
time, `match` has been implemented by the function `match-where`, which
is defined in terms of `match*`. This refactor moves the entire
definition into `match*`, removing the translation step.

Also, `match` now adds the `(or patt patts*)` pattern, which combines
multiple sub-patterns into one, matching if any sub-pattern matches.
`match` now refuses to compile patterns like `[(x)]`, which uses a
multival patterns in a place where only one item can match. Importantly,
the `or` pattern *does* support multivals, as in `(or (1 2) (2 1))`.

`or` patterns, like multivals, are not allowed to be nested deep inside
a pattern, and must appear at/near the top level.

There are two different ways that `or` patterns can expand. There is a
simple expansion, which takes place when there are no shared identifiers
introducing new variables to be bound. For example, the pattern (or 1 2)
compiles into two equality checks: (or (= _1_ 1) (= _1_ 2)).

The other expansion is more complicated, and happens if there are shared
identifiers. For example, the pattern (or {:foo x} {:bar x}) compiles
into the complicated code:
```fnl
(let [(matched?_2_ x_3_)
      (if (and (= (_G.type _1_) :table) (not= nil (. _1_ :foo)))
          (let [x (. _1_ :foo)] (values true x))
          (and (= (_G.type _1_) :table) (not= nil (. _1_ :bar)))
          (let [x (. _1_ :bar)] (values true x)))]
  (if matched?_2_ ...))
```
The emitted code checks each branch for a match, records the bindings,
and then potentially runs the body afterward.

After these changes, the following pattern shapes are allowed at the
top level:
`(where patt guards*)`
`(where (or patt patts*) guards*)`
`(or patt patts*)`

The following are not allowed:
`(where (where patt guards*) guards*)` -- The `where`s can be combinded
`(or (or pat patts*) patts)`           -- the `or`s can be combined
`(or (where patt guards*) patts*)`

It is possible to add support for that last pattern shape in the future.

Additionally, there is a new `matchless` macro, which behaves like
`match`, except it does not unify with other variables that are in
scope.