Responses to the suggestions to better-match
Renamed some variables and replaced some loops with their `icollect`
equivalents.

> * Replace without-multival with just a without function that takes a
>   key to remove.

The `without` function uses the `collect` macro, which was unfortunately
defined in terms of match. Changed `collect` to use `if`, so it should be
safe to invoke in `match.fnl` without self-definition problems.

> * Replace :multival? with :top? to indicate when it's OK to accept a
>   multival and when to disallow it.

I don't like this change, because the difference between
`top-level?` and `opts.top?` is unclear to a reader of the code.

To explain these flags:
They can't be merged into just one flag, because they are serving
different purposes. The `top-level?` flag permits `where` and `or`, and
the `opts.top?` flag permits multival.

If `where`, `or`, and multival are all in the same pattern, they need to
appear in this order:
```fnl
(where (or (my multival)))
```

To enforce this, there has to be a state at (where (or <>)) that shows
that <> can be multival, but can't be `where`/`or`. Right now, that
state is `top-level? = false`, and `opts.top? = true`.

I've left it alone for now, but I want to discuss / address this before
anything gets merged, because it's very unclear right now.

> * Factor out pre-binding-tail function into its own thing because the
>   match-condition function was getting kind of long and hard to
>   follow; even so I'm still not 100% sure what pre-bindings is for but
>   factoring out a separate function makes it a little easier to follow
>   I think? Not sure; maybe just some more comments would be enough.

Should now pass the tests again, and is hopefully clearer than the
original code. The `var` is back, as it is necessary, but it's now
factored into an `faccumulate`.

One thing to think about in the future:
The special cases in add-pre-bindings could be a lot cleaner if they
were moved into the `if*` special. Here's the cases that can be moved
to `if*`: make `(if true <body>)` emit the same lua as `<body>` and make
`(if <a> <b> <c> <d> true <body>)` emit `(if <a> <b> <c> <d> <body>)`.

If `if*` had these optimizations, the entire add-pre-bindings function
could be eliminated without hurting the quality of the emitted code.