Make case behave like discussed in the IRC
Unification only happens in match, or in a pattern with (=).
(=) is only allowed inside of a (where) pattern, and only inside case.
(or) is only allowed inside of a (where) pattern.

Basically, without a (where), the pattern must look like destructure
syntax, but inside of a (where), (or) and (=) can be used. The (where)
is required to be a visual reminder that the things inside the
pattern aren't calls to the usual `or` and `=` specials, but have
different pattern DSL meanings.