Disallow legacy ? guard in case
Instead of (a ? b) patterns, use (where a b) patterns inside of the case
special. For now, having a ? identifier as the second argument is an
error, but maybe in the future this can be changed to act as a multival
destructure, and completely remove the special case for ? in case.