Allow fennel.view to skip or apply options only once for the next call
This change brings more flexibility to the fennel.view function.  Now
it is possible to set options to "once" or "skip", which would mean
that for the current level, these options would be equal to true and
false respectively, but for nested levels, they will use their default
values.

(view "a" {:prefer-colon? {:once true}})
":a"

This means, that for this exact string `:prefer-colon?` is `true`.
However, if we were to use this with a table, it wouldn't work,
because for the table this option is not used, and for table elements,
this option will already use the default value of `false`:

(view ["a"] {:prefer-colon? {:once true}})
"[\"a\"]"

However, there's another syntax for this feature.

You can use a vector, which specifies how to treat the variable, and
what to set it to afterward. So, to make the previous example work, we
can use `{:after {:once true}}`. This will use `:once` for the first table,
which will not affect the table itself, but instead of setting the
option to the default value, it will set it to `:once` again:

(view ["a" ["b"] "c"] {:prefer-colon? {:once true :after {:once true}}})
"[:a [\"b\"] :c]"

Now the first level of the table is affected. This can be repeated to
propagate settings deeper `{:once true :after {:once true :after
{:once true}}}` if needed.

This syntax is also quite useful when working with the __fennelview
metamethod when you want to force your value on a specific data
structure, but use user-provided values in the nested ones:

{:__fennelview
 (fn [t view opts indent]
   (let [opts (doto opts
                (tset :prefer-colon?
                      {:once true
                       :after opts.prefer-colon?}))]
     (icollect [_ v (ipairs t)]
       (view v opts (+ 1 indent) true))))}

This will force colonized view of the strings inside this data
structure, but will preserve user-specified settings for nested data
structures.

It is also useful if you want to alter the behavior of existing
__fennelview metamethod with some options, like:

(fn vector [...]
  (->> {:__fennelview
        (fn [t view opts indent]
          (view t (doto opts
                    (tset :metamethod?
                          {:once false :after opts.metamethod?})
                    (tset :empty-as-sequence?
                          {:once true :after opts.empty-as-sequence?}))
                indent))}
       (setmetatable [...])))

(vector) => []
(vector []) => [{}]
(vector (vector)) => [[]]