stop readline default histfile ~/.rl_lua_history
Resolves https://todo.sr.ht/~technomancy/fennel/148

Earlier commit 84d28a2aa5b9656e27728b98a0db8a380913baf5 made is to that,
when lua-readline is present and a custom "histfile" not set, because
fennelrc was being invoked *before* setting the readline defaults, it
was falling back to using lua-readline's internal histfile of
~/.rl_lua_history.

This also meant persistent history was enabled, rather than disabled, by
default. This on its own is worth considering, but we shouldn't be
clobbering the default lua-readline history file regardless, since it
isn't a lua history :)

To fix this, the loader for fennelrc is being passed in as a REPL
option so it can be invoked *after* `try-readline`.