Reintroduce gensyms to unique renaming machinery
This fixes https://todo.sr.ht/~technomancy/fennel/54
in a different way than 6db6a2c that does not introduce
this bug: https://github.com/bakpakin/Fennel/issues/355

Approach: decomplect registering a demangling from extracting
a gensym's base. This way, the first time a gensym is
declared, it is not further mangled.

Note that neither commit completely addresses
https://github.com/bakpakin/Fennel/issues/355
e.g., consider (let [a 1 a 2] (lua "a"))
but this commit attempts to do an equivalent job
to 6db6a2c.