make lambdas check for metadata kv-table literals
Fixes https://todo.sr.ht/~technomancy/fennel/162

Lambdas weren't checking for kv table literals in metadata position.

As a result, when a lambda had a metadata literal and there were
arity-checking assertions injected, the metadata lost its position in
the body, so it was never recognized as metadata and left in the body of
the function.

```fennel
(macrodebug (? foo [a ...]
              {:fnl/arglist [a ?b c]}
              (let [(b c) (if (< (select :# ...) 2)
                            (values :B b)
                            ...)]
                [a b c])))
;; => (fn foo [a ...]
;;      (_G.assert (not= nil a) "Missing argument a on unknown:1")
;;      {:fnl/arglist [a ?b c]}
;;      (let [(b c) (if (< (select "#" ...) 2)
;;                    (values :B b)
;;                    ...)]
;;        [a b c]))
```

If the metadata included `:fnl/arglist`, this led to a false `unknown
identifier` compiler error.