make REPL more dynamic by allowing redefining its methods at runtime
This change makes it possible to redefine the currently running REPL
methods without restarting the REPL or starting a new one inside of
the current one. It is very useful for upgrading the REPL with new
features, and I think it will also be useful for solving the
https://todo.sr.ht/~technomancy/fennel/158 ticket.

The change is to provide all REPL methods as a ___repl___ table,
available in the REPL's environment, and to do lookups before invoking
each method, so the inner mechanism always gets the updated versions
of the methods.

What this patch provides can be summarized as the following code
snippet executed in the REPL:

>> (let [originalReadChunk ___repl___.readChunk]
     (fn ___repl___.readChunk [parser-state]
       (let [data (originalReadChunk parser-state)]
         (if (= data ",revert\n")
             (do (set ___repl___.readChunk originalReadChunk) "\n")
             (do (print (: "log: %s" :format data)) data)))))
>> (+ 1 2 3)
log: (+ 1 2 3)
6
>> ,revert
>> (+ 1 2 3)
6
>>

The readChunk was decorated to include logging of each expression
returned by readChunk, and a command was added to revert back to the
original implementation.

This patch also fixes a small bug in the fennel.view - detecting
cycles in tables with custom __pairs metamethod now works as well.