Allow macros to shadow specials and locals to shadow macros.
Up till this point, having a macro or special defined for a name would
prevent any locals from using the same name. This became a problem
when we introduced a new built-in macro named `case`, which was a
common enough name that it caused incompatibility with existing code
that used that name for a local. In addition, we want to allow macros
to shadow special forms, so that people can write their own version of
`fn`, etc.

However, we do not allow locals to shadow specials, tho we may add
that in the future. Currently we do not have a use case for this, and
adding new specials is much rarer than adding new macros.

We have several places where we look up identifiers in the scope table
to resolve them, and each look them up in different ways. First we
macroexpand the entire AST, which causes many identifiers to be
resolved in the scope.macros table. But the scope.macro table
currently looks in its parent if it's not found using a metatable;
this resolution chain completely ignores locals.

So for `find-macro` specifically, we need to use `rawget` to check
only the current scope and not its parents, and only keep looking up
the parent chain once we've confirmed that there is no local with that
name; the presence of a local will mean than `find-macro` must return
nil.

The `find-special` function is similar, except that specials can only
be defined in the root scope, so the presence of a macro anywhere in
the scope chain is enough to shadow the special and ensure it is not
resolved.

When looking up a local in `symbol-to-expression`, we first call
`find-macro` and if anything is found, we do not allow it to be
resolved as a local.

Finally, scopes do not always nest; for instance:

    (do (import-macros t :m)
        (local t {:hello #9})
        (t.hello))

In this case, the second definition of t must replace the original,
but they both are being defined in the same scope table. (And vice
versa if the local comes first and the macro second.) Because of this,
defining a local must wipe out the name from the macros table of the
current scope, and vice versa.