Consolidate stable-output tests in test/core.fnl
Originally, this was separated into its own module so it could run
compilation with the CLI. The reason for this was because order
instability in Lua < 5.3 is *random*, meaning it could sometimes
accidentally pass the test.

However, as it currently stands, this test consistently fails in the
main branch due to the string coercion part of the test cases. Further,
adding more keys to a single table allows a higher chance of variation
for the unstable order, and `make ci` will run the suite twice anyway:
once for Lua 5.3, and another time for Lua 5.4.

Since tests are slow, opting to skip use of the CLI and just run it in
the core test suite.