Allow metadata tables after docstring
With this change, it will be possible to follow the usual docstring
with metadata table literal, avoiding the need to nest the docstring
in the table, making it more convenient for both reading and writing
the code.

Additionally, metadata:get call will now return the whole metadata
table if no key is passed to it.