Add error when using vararg alone with an operator.
It is almost always a mistake to call (+ ...); people do this when
they assume that it will expand out to all the multivals, but we can't
do this without massive performance penalties. A compiler error draws
attention to the fact that their assumption is broken. Even tho the
code would technically work, there is no reason to write it this way.