fix case incorrectly binding = symbol
If `case` is used using (where (or)) to combine multiple patterns into
one big pattern, it is possible that a single symbol could be bound in
two different places.

```fnl
(case [1 2]
  (where (or {: x : y : z} [x y]))
  (+ x y))
```

In this example, `x` and `y` are introduced in *both* of the patterns,
and therefore they can be used in the binding. However, the indentifier
`z` only appears in one of the patterns; not both. For identifiers that
aren't in every subpattern, case does not attempt to bind the
identifier, so `z` cannot be used in the body.

The bug is caused with a special type of pattern that uses the = sign.
A pattern like `(= identifier)` can be used to match with a variable
that is already in scope. However, when `case` tries to compare multiple
patterns to look for shared identifiers, it incorrectly considered `=`
to be a valid identifier.

```fnl
(let [a 1
      b 2]
  (case 1
    (where (or (= a) (= b)))
    :good))
```

When analyzing this code, `case` treats `=` as an identifier that
should be bound in the body, similar to the way `x` and `y` were bound
in the previous example. However, `=` is not a valid name for a binding,
so the code erroneously fails to compile.

I fix the problem in the function `symbols-in-pattern`, which computes a
list of symbols a pattern wants to bind. It now is slightly smarter
about which parts of a pattern can introduce bindings.