generate man pages for tutorial, reference, api
Also moves fennel.1 to man/man1/fennel.1
The generated man pages are checked into the repo, so that they can be
installed with `make install` without users/packagers needing pandoc on
their machines.

This adds a pandoc filter for building manpages from the documentation.
Below are some details on the behavior of the custom pandoc filter:

* section: from extension of outfile e.g. fennel-api.3 > 3
* header (sets top center of man page): set from first top-level header
* title (sets top left corner of manpage): from base of outfile, e.g.
  man/man7/fennel-tutorial.7 -> fennel-tutorial
* date: Use current date if one isn't supplied by --metadata date=foo

Formatting and output:
* Force uppercase for top-level headers
* Promote all headers up by 1 (unless it's already 1)
* Format table output to be manpage-friendly - pandoc doesn't support
  this by default