Optimize compiler output for "if true"
When the "else" branch is missing and the last condition is the literal
`true`, optimize out the condition.

For example, (if true x) gets converted to (do x), and
(if x y true z) gets converted to (if x y z)

This optimization does not apply when the "else" branch is present.

(if true x nil) is not touched.