manfilter: remove superfluous debug env var check
The pandoc filter for man page generation writes info to stderr about the metadata
being set for the resulting man page.

This debug function had been checking an environment variable flag for
disablign debug output, which was both unnecessary for this *and* had a
bug making it always enable debug regardless of the env var.

Rather than fixing it, I opted to simply remove it the flag until we
actually need it.