ensure (< x y z) protects wraps output in parentheses
This is to prevent edge cases where operator precedence issues change
the outer expressions:

This is a somewhat unrealistic example, but:
`(+ x (<= a b c))` previously compiled to `x + a <= b and b <= c`
The precedence of `+` is greater than `and`, so it evaluates `(x + a)`
bbefore checking the rest of the comparator.

The fix is as simple as wrapping the whole thing in parentheses.