handle os.execute return values in cross-lua way
The previous implementation wasn't taking into account the differences
in lua versions:
* In Lua 5.2+, os.execute returns (values ok reason code), where reason
  is either "exit" or "signal" and code is either exit status or the
  signal that terminated it.
* In Lua 5.1, it returns only the code

This is a basic adjustment at least makes sure the success/failure
status is passed on. It doesn't attempt to actually pass the matching
code, as I'm not sure whether it would be desirable in the case of
terminating due to a signal to pass said signal's code to (os.exit).