Revert change to include modexpr calculation.
In 5c803e426a5fff59ace6ab7520d14be63cb93c6e this was moved out of the
modexpr let; however, this appeared to have been a mistake. We need a
fallback if the module name can't be resolved normally, and there is
no reason to double-compile this if the module name did get resolved
already.