Prevent metadata table from being accessible in compiler sandbox.
The problem here is that the *keys* of the metadata table are
functions. Normally they're nothing too dangerous (fennel.eval is not
in the table) but if you're running with metadata enabled, your
application's own code could have functions in this table, and you
really don't want all that exposed to the compiler sandbox.