Add pre- variants for several hooks.
* pre-do
* pre-fn
* pre-each
* pre-for

These are largely based on the equivalent customhook-early-fn hooks in
the fork of Fennel used by fennel-ls, as they are needed to get
additional scope information for these various forms. However, other
than the name, they differ in a few ways.

They try to maintain the current convention of putting the scope table
second instead of using it as the final argument.

The :pre-each hook provides the bindings, the iterator, and optional
until condition.