Don't reset parser's last byte after every repl loop.
When the parser encounters something like this: 1)

It parses 1 and then leaves ) in the lastb var so that it would get
picked up by the next getb call working thru the stream. However, when
the repl calls parse, it would call `reset` after every loop in order
to set the line/byte/column back for each fresh input string.

Previously calling `reset` would also reset lastb, meaning the
spurious input would get discarded between repl loops. During normal
repl runs, lastb is a newline, so discarding it is fine. But in the
case of trailing characters after a number/symbol/boolean, the
leftover character would get discarded.

With this change, only discard `lastb` when it's a newline; otherwise
keep it around so we can detect bad input.