`fennel-module-name` compiler-env helper
There are a few places where we've internally referenced the active
fennel module name for dynamically requiring fennel during runtime,
accounting for a non-standard module name. This adds a
`fennel-module-name` helper to the compiler env to serve this purpose,
allowing it to be accessed from macros.

Also updated `assert-repl` (which prompted this change), function
metadata insertion, and `with-open` to use the helper.

This change theoretically opens the path to userland macros
conditionally accessing the fennel API even when it's been loaded via an
alternate module path, but we'd require further changes to make it
usable, as the fennel module name isn't automatically detected and is
assumed to be `fennel` unless the undocumented `moduleName` option
(which should stay undocumented to avoid confusion with the
also-undocumented 'module-name') is passed.

Holding off on documenting this for now until we can ensure fully
supported usage, with automatic detection of module name when fennel is
required.