Remove test assertion for assert-repl interactive message.
An interactive assert-repl doesn't propagate the original failure
message and it doesn't let you set the failure message when you
,return (values false "oh no") because the return repl command just
doesn't bother with multivals yet.

This test was passing by accident in PUC because the error message
included the input string, not because the actual correct failure
message was being propagated. But it failed in LuaJIT.