catch more unexpected paren cases in iter binding
Needed to reorg where the iterator pragmas/clauses are checked, since
when occurred in the destructure phase, an invalid :foo/&foo clause is
misinterpreted as the sourcei iterator, wrongly triggering an
"unexpected iterator bindings" compiler error if the source iterator is
a list.

Also updated utils.string? to return the string instead of `true` on
success, which makes it a little more useful.