Allow -?> and -?>> to skip gensym
This patch changes the output of -?> and -?>> to only bind their
argument to an identifier if it isn't already idempotent. To do this,
the (double-eval-safe?) helper function is promoted to the top of
src/fennel/macros.fnl, so the logic can be cleanly shared across the
macros.

Each macro is written so that it expects its argument to be idempotent.
Whenever this expectation is broken, the macro simply binds the argument
and recursively re-invokes itself.

The doto macro also had logic to prevent unnecessary bindings, and this
patch refactors that logic so that it is in the same format as the -?>
and -?>> macros.