Fix multival returns in -?> and -?>>
My previous patch messed up the -?> macro. Although -?> and -?>> don't
handle intermediate multivals, their final step can be a multival, and
shouldn't be stored in a variable.

The (doto) macro also was changed in the previous patch, but doesn't
need to be fixed; it still works as intended.

There's a careful balance between making `-?>` a recursive function, and
making `-?>` emit a call to `-?>` that needs to go through another
macroexpand step. A recursive call will run much faster, but it will also
skip the autogensym step, so all of the `tmp#` identifiers would receive
the same symbol. Because of this, the `tmp#` branch emits a `-?>` call in
the output to force autogensym, but the other call is native recursion
for speed.