Use git submodules for binary builds instead of tarballs.
Previously we were using curl to fetch tarballs from lua.org because
the unofficial github mirror of lua is not very good. But I realized
we can use a submodule of the debian packaging for lua, which is much
better.

This removes one bootstrapping dependency (curl) but also moves more
of our dependencies into git-land where they are fully reproducible
and content-addressed.

As a bonus, this makes it so that `make fennel-bin LUA=lua/src/lua`
should work even on a system that doesn't have any lua present.