Disallow introducing locals when passing args to functions/macros.
We can't prevent functions from being declared with names, because the
name might be needed for recursion. But we can detect and declare the
local name in the function's internal scope instead of the containing
scope where appropriate.

Except we're currently doing both of these based on nval=1, which
means we can't prevent new names from being declared in the last
argument to a function.