Add comment listing macros.fnl compiler-env extras
Built-ins in src/fennel/macros.fnl are compiled using a custom
compiler-env exposing fennel, utils, and specials.get-function-metadta,
but it's easy to forget these are available that way.

A better solution may be to make it possible to explicitly require the
relevant internal modules from macros.fnl, but a comment is a better
than nothing.